# Improve location efficiency